Move lti_launch.py to _lti_launch.py
The module `lms/util/lti_launch.py` has the exact same name as one of the
functions it contains, `lti_launch()`.

`lms/util/__init__.py` imports the `lti_launch()` function
(`from lms.util.lti_launch import lti_launch`). This means that if tests
for another module try to patch the `lti_launch.py` module with
`patch("lms.util.lti_launch")` it will patch the `lti_launch()` function
not the `lti_launch.py` module. The module can't be patched.

Fix this by renaming the module to `_lti_launch.py`, which also
clarifies that it's a private internal module of `lms.util` that other
packages shouldn't import directly. The `_lti_launch.py` functions that
other code uses are already exported in `lms/util/__init__.py` as
top-level parts of `lms.util`'s public interface:
`lms.util.get_application_instance()` and `lms.util.lti_launch()`.
