+++ lms/util/__init__.py
@@ -4,8 +4,8 @@ from lms.util.authenticate import authenticate
 from lms.util.authorize_lms import authorize_lms, save_token
 from lms.util.canvas_api import canvas_api, GET, POST
 from lms.util.jwt import jwt
-from lms.util.lti_launch import get_application_instance
-from lms.util.lti_launch import lti_launch
+from lms.util._lti_launch import get_application_instance
+from lms.util._lti_launch import lti_launch
 from lms.util.lti import lti_params_for
 from lms.util.via import via_url
 from lms.util.view_renderer import view_renderer
